[JBPM-9255] Fix TaskEvent.processType for validation at some dbs

+++ jbpm-db-scripts/src/main/resources/db/ddl-scripts/sqlserver2008/sqlserver2008-jbpm-schema.sql
@@ -414,7 +414,7 @@
         OPTLOCK int,
         workItemId bigint,
         correlationKey varchar(255),
-        processType smallint,
+        processType int,
         primary key (id)
     );
